Refactor: New compnents and utiliti functions
