+++ app/component/itinerary/NaviStack.js
@@ -1,53 +1,21 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { intlShape } from 'react-intl';
 import cx from 'classnames';
-import { configShape } from '../../util/shapes';
-import Icon from '../Icon';
+import NaviMessage from './NaviMessage';
 
 // eslint-disable-next-line no-unused-vars
-const NaviStack = ({ notifications, handleRemove, show }, { intl, config }) => {
-  const [removingIndex, setRemovingIndex] = useState(null);
-
-  const handleRemoveClick = index => {
-    setRemovingIndex(index);
-    setTimeout(() => {
-      handleRemove(index);
-      setRemovingIndex(null);
-    }, 500);
-  };
+const NaviStack = ({ notifications, handleRemove, show }) => {
   return (
     <div className={cx('info-stack', !show ? 'slide-out' : 'slide-in')}>
-      {notifications.map((info, index) => (
-        <div
-          key={info.backgroundColor}
-          style={{ backgroundColor: info.backgroundColor }}
-          className={cx(
-            'info-stack-item',
-            removingIndex === index ? 'slide-out-right' : '',
-          )}
+      {notifications.map((notification, index) => (
+        <NaviMessage
+          key={notification.id}
+          severity={notification.severity}
+          index={index}
+          handleRemove={handleRemove}
         >
-          <Icon
-            img={info.iconId}
-            height={1.4}
-            width={1.4}
-            className="info-icon"
-            color={info.iconColor}
-          />
-          {info.content}
-          <button
-            type="button"
-            className="info-close"
-            onClick={() => handleRemoveClick(index)}
-          >
-            <Icon
-              img="icon-icon_close"
-              height={0.8}
-              width={0.8}
-              color={config.colors.primary}
-            />
-          </button>
-        </div>
+          {notification.content}
+        </NaviMessage>
       ))}
     </div>
   );
@@ -57,10 +25,5 @@ NaviStack.propTypes = {
   notifications: PropTypes.arrayOf(PropTypes.object).isRequired,
   show: PropTypes.bool.isRequired,
   handleRemove: PropTypes.func.isRequired,
-};
-
-NaviStack.contextTypes = {
-  intl: intlShape.isRequired,
-  config: configShape.isRequired,
 };
 export default NaviStack;
